Use JWE to pass forwarded URL
We can use the `state` parameter required by the GitHub OAuth flow to
pass the forwarded URL as an encrypted JSON web token. This means we
can redirect the user directly to the GitHub OAuth endpoint when their
request arrives at the route service.

On a successful authentication, we need to redirect the user twice in
order to provide the destination service with a cookie containing the
GitHub access token. The first redirect is from the callback to the
original requested endpoint with the encrypted GitHub access token in a
query parameter. The next redirect sets the token cookie and removes the
token query parameter.

This route service can be bound to multiple backing services that
require membership to a specific GitHub organisation.